Move fileobject type definition to nameable complexType
xmllint fails to validate a schema where an embedded complexType
definition has a name attribute.  There is at least one reason one
would want to specifically reference the type definition of a
fileobject---embedding a copy of the file object, as in the (soon to be
proposed) differential annotations---so this definition is exported to
its own, name-able element.

Signed-off-by: Alex Nelson <a.nelson@prometheuscomputing.com>