Allow for extensibility in commonly-changing elements
The elements and attributes extended in this commit are the minimum
needed for differential annotations as idifference.py currently
generates, also allowing for other extensions on volumes and
fileobjects.

As written, the DFXML schema allowed for no extensibility of elements,
including from new namespaces.  To allow a valid .dfxml file to include
non-"dfxml:" namespaced elements and attributes, the `<any>` element
must be included in the `<element>` definition.

To illustrate what this patch allows: If a modification timestamp is
changed on a file, to express that change in differentially-
annotated DFXML (under the "delta" namespace), idifference.py adds
these annotations:

    <fileobject delta:changed_properties="1">
      ...
      <mtime delta:changed_property="1">2013-09-19:01:23:45Z</mtime>
      ...
    </fileobject>

With the changes in this patch, dfxml.xsd will validate idifference.py's
extended-DFXML output.  To see for yourself, issuing these commands:

    git clone --branch=delta_schema https://github.com/ajnelson/dfxml.git
    cd dfxml/python
    make check-validate

will ultimately give you this output:

    xmllint --noout --schema ../schema/delta.xsd idifference_test_formatted.dfxml
    idifference_test_formatted.dfxml validates
    xmllint --noout --schema ../deps/dfxml_schema/dfxml.xsd idifference_test_formatted.dfxml
    idifference_test_formatted.dfxml validates

Without the changes in this patch, any added element causes validation
to fail.  This goes against the spirit of XML extensibility.

A downside this patch illustrates is that every element and attribute
must be permitted to be extended.  Suggestions for alternative extension
mechanisms are welcome.

Signed-off-by: Alex Nelson <a.nelson@prometheuscomputing.com>